adds toggleElements function to show and hide form, fetching message, and score of movie
