+++ backend/model/redditScraper.js
@@ -19,29 +19,20 @@ scraper = {
       })
       postTitles.map(title => allPostTitles.push(title));
 
-
-
       let nextPage = await page.evaluate(() => {
-        console.log('page being evaluated for next button')
 
         if (document.querySelector('.nextprev > a[rel~=next]')) {
-          console.log('next button');
           return document.querySelector('.nextprev > a[rel~=next]').href;
         } else {
-          console.log('no next button');
           return false;
         }
       })
 
       if (nextPage) {
-        console.log(`next page ${nextPage}`);
-
         await page.goto(nextPage, {waitUntil: 'load', timeout: 0});
-
         //recursive call to scrape the next page
         await scrapeCurrentPage();
       } else {
-
         return;
       }
     }
@@ -50,16 +41,10 @@ scraper = {
     let page = await browser.newPage();
 
     await page.goto(this.url, {waitUntil: 'load', timeout: 0});
-
     //wait until search box loads
     await page.waitFor('form#search');
-
     //type in movie title
     await page.click('input[name=q]');
-    // let moviesToReview = ['Wet Hot American Summer'];
-    // let chosenMovie = moviesToReview[Math.floor(Math.random() * moviesToReview.length)];
-
-
     await page.$eval('input[name=q]', (el, movie) =>  {
       el.value = `${movie}`
     }, movie)
@@ -71,9 +56,7 @@ scraper = {
     await page.waitForNavigation()
     await scrapeCurrentPage();
 
-    // return [allPostTitles.join(''), chosenMovie]
     return allPostTitles.join('');
-
   }
 }
